+++ akshare/__init__.py
@@ -1899,9 +1899,10 @@ amac_manager_cancelled_info # 中国证券投资基金业协会-信息公示-诚
 1.4.76 add: add stock_lhb_stock_statistic_em interface
 1.4.77 fix: fix stock_hk_ggt_components_em interface
 1.4.78 add: add remove matplotlib module and support cache
+1.4.79 add: add stock_board_industry_hist_min_em interface
 """
 
-__version__ = "1.4.78"
+__version__ = "1.4.79"
 __author__ = "AKFamily"
 
 import sys
@@ -2053,6 +2054,7 @@ from akshare.stock.stock_repurchase_em import stock_repurchase_em
 from akshare.stock.stock_board_industry_em import (
     stock_board_industry_cons_em,
     stock_board_industry_hist_em,
+    stock_board_industry_hist_min_em,
     stock_board_industry_name_em,
 )
 
@@ -2334,6 +2336,7 @@ from akshare.economic.macro_swiss import (
 from akshare.stock.stock_board_concept_em import (
     stock_board_concept_cons_em,
     stock_board_concept_hist_em,
+    stock_board_concept_hist_min_em,
     stock_board_concept_name_em,
 )