+++ akshare/stock/stock_board_concept_em.py
@@ -165,6 +165,77 @@ def stock_board_concept_hist_em(symbol: str = "数字货币", adjust: str = "") 
     return temp_df
 
 
+def stock_board_concept_hist_min_em(symbol: str = "长寿药", period: str = "5") -> pd.DataFrame:
+    """
+    东方财富网-沪深板块-概念板块-分时历史行情
+    http://quote.eastmoney.com/bk/90.BK0715.html
+    :param symbol: 板块名称
+    :type symbol: str
+    :param period: choice of {"1", "5", "15", "30", "60"}
+    :type period: str
+    :return: 分时历史行情
+    :rtype: pandas.DataFrame
+    """
+    stock_board_concept_em_map = stock_board_concept_name_em()
+    stock_board_code = stock_board_concept_em_map[
+        stock_board_concept_em_map["板块名称"] == symbol
+    ]["板块代码"].values[0]
+    url = "http://91.push2his.eastmoney.com/api/qt/stock/kline/get"
+    params = {
+        "secid": f"90.{stock_board_code}",
+        "ut": "fa5fd1943c7b386f172d6893dbfba10b",
+        'fields1': 'f1,f2,f3,f4,f5,f6',
+        'fields2': 'f51,f52,f53,f54,f55,f56,f57,f58,f59,f60,f61',
+        'klt': period,
+        'fqt': '1',
+        'end': '20500101',
+        'lmt': '1000000',
+        '_': '1647760607065',
+    }
+    r = requests.get(url, params=params)
+    data_json = r.json()
+    temp_df = pd.DataFrame([item.split(",") for item in data_json["data"]["klines"]])
+    temp_df.columns = [
+        "日期时间",
+        "开盘",
+        "收盘",
+        "最高",
+        "最低",
+        "成交量",
+        "成交额",
+        "振幅",
+        "涨跌幅",
+        "涨跌额",
+        "换手率",
+    ]
+    temp_df = temp_df[
+        [
+            "日期时间",
+            "开盘",
+            "收盘",
+            "最高",
+            "最低",
+            "涨跌幅",
+            "涨跌额",
+            "成交量",
+            "成交额",
+            "振幅",
+            "换手率",
+        ]
+    ]
+    temp_df["开盘"] = pd.to_numeric(temp_df["开盘"])
+    temp_df["收盘"] = pd.to_numeric(temp_df["收盘"])
+    temp_df["最高"] = pd.to_numeric(temp_df["最高"])
+    temp_df["最低"] = pd.to_numeric(temp_df["最低"])
+    temp_df["涨跌幅"] = pd.to_numeric(temp_df["涨跌幅"])
+    temp_df["涨跌额"] = pd.to_numeric(temp_df["涨跌额"])
+    temp_df["成交量"] = pd.to_numeric(temp_df["成交量"])
+    temp_df["成交额"] = pd.to_numeric(temp_df["成交额"])
+    temp_df["振幅"] = pd.to_numeric(temp_df["振幅"])
+    temp_df["换手率"] = pd.to_numeric(temp_df["换手率"])
+    return temp_df
+
+
 def stock_board_concept_cons_em(symbol: str = "车联网") -> pd.DataFrame:
     """
     东方财富-沪深板块-概念板块-板块成份
@@ -277,6 +348,9 @@ if __name__ == "__main__":
         symbol="车联网", adjust=""
     )
     print(stock_board_concept_hist_em_df)
+
+    stock_board_concept_hist_min_em_df = stock_board_concept_hist_min_em(symbol="长寿药", period="5")
+    print(stock_board_concept_hist_min_em_df)
 
     stock_board_concept_cons_em_df = stock_board_concept_cons_em(symbol="网络安全")
     print(stock_board_concept_cons_em_df)